TicTacToeAI-O
AI which takes the data set from the generator function, and uses it to fit a deep learning classifier to predict the best possible move for both X and O in any condition. the script then uses the O classifier to play against the user and never lose. The output is shown as a simple UI in the terminal as a numerical matrix with 1 as X, -1 as O and 0 as a free square. (To play as O, change line 101 to "if turn = false" and change line 191 to "probmatrix=xfullMLPC.predict_proba(testdata)" )